requirements.txt
dbt-postgres==1.9.0
soda-core-postgres==3.5.5
psycopg2-binary==2.9.10
python-dotenv==1.1.1
